Issue #1010: Move logging decorator factory method from common factory into the logging factories
